refactor(plugins): move google drive models into plugin
