+++ app/core_plugins/googledrive/tests/test_routes.py
@@ -11,8 +11,8 @@ from httpx import AsyncClient
 from sqlmodel import select
 from sqlmodel.ext.asyncio.session import AsyncSession
 
+from app.core_plugins.googledrive.models import DriveFile, DriveFolder, DriveOAuthToken
 from app.lib.documents import Document, DocumentStatus
-from app.models.drive import DriveFile, DriveFolder, DriveOAuthToken
 from app.models.user import User
 
 # ---------------------------------------------------------------------------
